requirements.txt
fastapi
uvicorn
requests
google-generativeai
python-dotenv
pydantic
twilio
